requirements.txt
aiogram==2.0
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
install==1.3.5
multidict==6.0.5
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.0.0
sqlparse==0.5.0
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
